Fix decryption failure when multiple cipher filter blocks are used
Threads are not tied to declared filter blocks in the config so the cipher instance stored in a particular thread may get events from any of the configured instances of the filter, which may not match the key that it is currently configured to use.  This simple fix does away with persisting the cipher instance between filter calls and always initializes a new instance at the start to ensure that the proper encryption key for the current event is used.